MAINT: store/load `state_dict` of model/optimizer into/from files
`state_dict` is a mutable object (literally, it a `dict`), so that
it can be changed in runtime.

In the original version, we just hold a reference of `state_dict`,
not a copied instance. So that it won't work correctly if we want
to reset the state after running `LRFinder.range_test()`.

In order to keep the original state of model and optimizer, we have
to:
  1. make a deep copy of `state_dict`
  2. save `state_dict` into file, then load it back later

We choose the second method. Although it take a little more time on
reading and writing file, it avoids consuming extra memory. So that
it won't lead to an out-of-memory issue.