UPDATE: add a new input parameter `memory_cache`
User now can select the strategy of caching `state_dict` of model
and optimizer by setting `memory_cache`.

By default, the value is `Ture` and it indicates that `state_dict`
will be cached in memory. If users need to run `range_test` with a
large model and prevent out-of-memory exception occuring, they can
set `memory_cache` to `False` to make `state_dict` to be saved in
file.