MAINT: store/load `state_dict` of model/optimizer into/from files (#1)
* MAINT: store/load `state_dict` of model/optimizer into/from files

`state_dict` is a mutable object (literally, it a `dict`), so that
it can be changed in runtime.

In the original version, we just hold a reference of `state_dict`,
not a copied instance. So that it won't work correctly if we want
to reset the state after running `LRFinder.range_test()`.

In order to keep the original state of model and optimizer, we have
to:
  1. make a deep copy of `state_dict`
  2. save `state_dict` into file, then load it back later

We choose the second method. Although it take a little more time on
reading and writing file, it avoids consuming extra memory. So that
it won't lead to an out-of-memory issue.

* MAINT: rewrite code for loading and saving model state

Those code for loading and saving model state are incorrect in
previous commit...

Besides, the indentation is also wrong in the section of
`LRFinder.reset()`.

* UPDATE: add a new input parameter `memory_cache`

User now can select the strategy of caching `state_dict` of model
and optimizer by setting `memory_cache`.

By default, the value is `Ture` and it indicates that `state_dict`
will be cached in memory. If users need to run `range_test` with a
large model and prevent out-of-memory exception occuring, they can
set `memory_cache` to `False` to make `state_dict` to be saved in
file.

* MAINT: make cached `state_dict` be reusable, and add a check before
`StateCacher` being destroyed

In order to make user able to reuse cached `state_dict` for running
`range_test()` with different settings, the method of getting cached
content in `StateCacher.retrieve` is changed from `cached.pop()` to
`cached.get()`.

Besides, a check is add in `StateCacher.__del__()` to make sure all
cached files will be removed before a `StateCacher` instance being
destroyed.