UPDATE: implement a new LRFinder with the support of gradient accumulation
`AccumulationLRFinder` is a learning rate finder implemented with the
mechanism of gradient accumulation.

Besides, the iterator used for getting batch of data for training is
now replaced by `DataLoaderIterWrapper` to simplify the code and make
the implementation of `AccumulationLRFinder` easiler.

And the input parameter ofs `LRFinder._train_batch()` is also modified
for the same reason mentioned above.