API: Add non_blocking_transfer argument to range_test
This parameter controls whether to use the non_blocking
flag when transferring data from the CPU to the CUDA device.

In general, it is fine to use a non_blocking transfer, but
the option to disable it is provided in case the user has
problems with very large tensors allocated to pinned memory.