MAINT: determine batch size from `labels` instead of `inputs`
In `LRFinder._validate()`, batch size was determined by given
`inputs` dynamically. However, `inputs` might not be tensors
or list of tensors in some cases, and that makes it not able to
determine batch size from `inputs.size()` or `inputs[0].size()`.

In this revision, we change to code to determine batch size from
`labels` since all loss functions expecting the `labels` to be a
tensor (reviewed by @davidtvs). And it keeps the compatibility
to work with non-full batch (e.g. when `drop_last` of a
`DataLoader` is True).