Fix issue #65
`min_grad_idx` is a numerical value but it was not properly used to
determined whether suggested lr should be returned through the
following statement:
```python
if suggest_lr and min_grad_idx:
```

As @manuel-munoz-aguirre mentioned in issue #65, this IF-statement
will fail to return suggested lr when `min_grad_idx` is 0.

A proper solution for this issue should be explicitly checking
whether `min_grad_idx` is `None` or not (as it was initialized to
`None` in line 503).